Applied changes to reflect requirements
